Implemented the matrix generator and the getFitting method

+++ src/EvolutionaryProgramming/Solution.java
@@ -1,7 +1,7 @@
 package EvolutionaryProgramming;
 
 public interface Solution {
-	public float getFitting();
+	public double getFitting();
 	
 	public void generateRandomSolution();
